fix(tests): stop the subscription-set suite depending on import order
CI aborted the whole 37-test SubscriptionSetProvider suite before its first
assertion with `ReferenceError: document is not defined`, thrown out of
goober while react-hot-toast was still being evaluated. The same suite passed
locally, which is the part worth explaining: bun:test shares one process
across files, an earlier suite had left a `document` global behind, and the
file order happened to differ. The tests were green by accident, and the
accident did not survive a different machine.

Splits the DOM-free half of the provider into subscriptionSetState.ts — the
reducers, the write classifiers, the anonymous/fallback states and the
sign-out signal — and points the suite at that. The module has no runtime
imports at all (its one import is type-only and erased), so no import order
can turn it green or red again.

The provider re-exports the same surface, so nothing else changes shape.
Navbar and SignedOutGate now reach for the state module directly instead,
which also keeps react-hot-toast out of two chunks that never toast anything.
